refactor(skills): extract FileTreeNode and SearchResultRow components
Extract two memoized components from the monolithic SkillsSettings.tsx
(1,262 lines) into dedicated files:

- FileTreeNode.tsx (95 lines) - recursive file tree node component
- SearchResultRow.tsx (117 lines) - search result row component

Also removed unused constants and styled components that were only
used by the extracted components.

SkillsSettings.tsx reduced from 1,262 to 1,070 lines.
